requirements.txt
# Core bot dependencies
aiogram==3.4.1
aiohttp==3.9.1
aiofiles==23.2.1
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0

# Redis (for rate limiting and caching)
redis==5.0.1
aioredis==2.0.1

# HTTP requests
httpx==0.26.0
requests==2.31.0

# Payment processing
stripe==7.12.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# GitHub integration
PyGithub==2.1.1

# Crypto utilities
bip-utils==2.9.3
cryptography==41.0.8

# Utilities
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
uvloop==0.19.0

# Development and testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring
prometheus-client==0.19.0
psutil==5.9.7

# Logging
loguru==0.7.2

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Date/time
python-dateutil==2.8.2
pytz==2023.3

# Environment
python-decouple==3.8
